requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pymysql==1.1.0
pymssql==2.3.1
psycopg2-binary==2.9.9
pandas==2.2.2
jinja2==3.1.4
python-dotenv==1.0.1
cryptography==42.0.8
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
openpyxl>=3.1.0
reportlab>=4.0.0
PyYAML>=6.0
